First step refactor
